+++ mosaic-testing/src/commonMain/kotlin/com/jakewharton/mosaic/testing/TestMosaic.kt
@@ -108,6 +108,7 @@ private class RealTestMosaic<T>(
 	override fun draw() = mosaic.draw()
 	override fun static() = mosaic.static()
 	override fun dumpNodes() = mosaic.dumpNodes()
+	override fun title() = mosaic.title()
 
 	override suspend fun awaitComplete() = mosaic.awaitComplete()
 	override fun cancel() = mosaic.cancel()